again, slow progress...

+++ BeeColonySimulator/partie7/src/Env/HelperFunctions.cpp
@@ -9,11 +9,11 @@ int help::getIndex(const Vec2d &p, float const& size, int const& nb) {
 }
 
 int help::getX(const Vec2d &p, const float &size) {
-    return trunc(p.x()/size);
+    return (int)trunc(p.x()/size);
 }
 
 int help::getY(const Vec2d &p, const float &size) {
-    return trunc(p.y()/size);
+    return (int)trunc(p.y()/size);
 }
 
 double help::temperatureFunction1(double temp) {
